+++ packages/widgetbook/lib/test.dart
@@ -1,2 +1,1 @@
-export 'src/core/accessibility/guidelines.dart';
 export 'src/test/test.dart' hide outputDir;